Inserting at End
This code is for inserting an element at the bottom of the stack. In this code we are taking input for the - Number of elements in the stack, Stack elements, Element to be inserted at the end. First we are applying a while loop and moving the stack elements in another Stack. Then we are inserting the desired element in the original stack which is now empty. Then we are applying the while loop again for moving the elements in the original stack back.